Refactor our two end-to-end tests a bit to remove some duplication
